+++ karabo_bridge/__init__.py
@@ -38,10 +38,10 @@ __version__ = "0.4.0"
 
 from .cli import *
 from .client import *
-from .serialize import *
+from .serializer import *
 from .simulation import *
 
 
 __all__ = (client.__all__ +
            simulation.__all__ +
-           serialize.__all__)
+           serializer.__all__)
